+++ assignment-2/app/src/main/java/com/example/assignment2/MyAdapter.kt
@@ -1,0 +1,24 @@
+package com.example.assignment2
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+
+import androidx.recyclerview.widget.RecyclerView
+import com.example.assignment2.databinding.DrawerBinding
+
+class MyAdapter(private val data: List<MainActivity.MyData>): RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    override fun getItemCount(): Int {
+        return data.size
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+        return MainActivity.DrawerViewHolder(DrawerBinding.inflate(LayoutInflater.from(parent.context)))
+    }
+
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+        val info = data[position]
+        holder as MainActivity.DrawerViewHolder
+        holder.bind(info)
+    }
+
+}
